feat: move http-client to @novu/client
