+++ packages/client/src/http-client/index.ts
@@ -1,1 +1,1 @@
-export * from './api.client';
+export * from './http-client';
